Merge Master branch from Main trunk

+++ org.adempiere.base/src/org/compiere/dbPort/Convert.java
@@ -37,7 +37,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.compiere.db.Database;
-import org.compiere.model.MSysConfig;
 import org.compiere.util.CLogger;
 import org.compiere.util.DisplayType;
 import org.compiere.util.Env;
@@ -294,23 +293,23 @@ public abstract class Convert
 	 * @param retVars
 	 * @return string
 	 */
-	protected String replaceQuotedStrings(String inputValue, Vector<String>retVars) {
+	protected String replaceQuotedStrings(String inputValue, Vector<String>retVars, String nonce) {
 		// save every value  
 		// Carlos Ruiz - globalqss - better matching regexp
 		retVars.clear();
 		
 		// First we need to replace double quotes to not be matched by regexp - Teo Sarca BF [3137355 ]
-		final String quoteMarker = "<--QUOTE"+System.currentTimeMillis()+"-->";
+		final String quoteMarker = "<--QUOTE"+nonce+"-->";
 		inputValue = inputValue.replace("''", quoteMarker);
 		
 		Pattern p = Pattern.compile("'[[^']*]*'");
 		Matcher m = p.matcher(inputValue);
 		int i = 0;
-		StringBuffer retValue = new StringBuffer(inputValue.length());
+		StringBuilder retValue = new StringBuilder(inputValue.length());
 		while (m.find()) {
 			String var = inputValue.substring(m.start(), m.end()).replace(quoteMarker, "''"); // Put back quotes, if any
 			retVars.addElement(var);
-			m.appendReplacement(retValue, "<--" + i + "-->");
+			m.appendReplacement(retValue, "<--QS" + i + "QS" + nonce + "-->");
 			i++;
 		}
 		m.appendTail(retValue);
@@ -325,12 +324,12 @@ public abstract class Convert
 	 * @param retVars
 	 * @return string
 	 */
-	protected String recoverQuotedStrings(String retValue, Vector<String>retVars) {
+	protected String recoverQuotedStrings(String retValue, Vector<String>retVars, String nonce) {
 		for (int i = 0; i < retVars.size(); i++) {
 			//hengsin, special character in replacement can cause exception
 			String replacement = (String) retVars.get(i);
 			replacement = escapeQuotedString(replacement);
-			retValue = retValue.replace("<--" + i + "-->", replacement);
+			retValue = retValue.replace("<--QS" + i + "QS" + nonce + "-->", replacement);
 		}
 		return retValue;
 	}
@@ -385,7 +384,7 @@ public abstract class Convert
 	
 				} catch (Exception e) {
 					String error = "Error expression: " + regex + " - " + e;
-					log.info(error);
+					log.warning(error);
 					m_conversionError = error;
 				}
 			}
@@ -515,6 +514,7 @@ public abstract class Convert
 			"AD_WINDOW_ACCESS",
 			"AD_WORKFLOW_ACCESS",
 			"AD_WORKFLOWPROCESSORLOG",
+			"AD_USERPREFERENCE",
 			"CM_WEBACCESSLOG",
 			"C_ACCTPROCESSORLOG",
 			"K_INDEXLOG",
@@ -549,6 +549,9 @@ public abstract class Convert
 			return true;
 		if (uppStmt.startsWith("UPDATE R_REQUESTPROCESSOR SET DATELASTRUN"))
 			return true;
+		// don't log sequence updates
+		if (uppStmt.startsWith("UPDATE AD_SEQUENCE SET CURRENTNEXT"))
+			return true;
 		// Don't log DELETE FROM Some_Table WHERE AD_Table_ID=? AND Record_ID=?
 		if (uppStmt.startsWith("DELETE FROM ") && uppStmt.endsWith(" WHERE AD_TABLE_ID=? AND RECORD_ID=?"))
 			return true;
@@ -581,13 +584,8 @@ public abstract class Convert
 
 	private static void writeLogMigrationScript(Writer w, String statement) throws IOException
 	{
-		boolean isUseCentralizedID = "Y".equals(MSysConfig.getValue(MSysConfig.DICTIONARY_ID_USE_CENTRALIZED_ID, "Y")); // defaults to Y
-		boolean isUseProjectCentralizedID = "Y".equals(MSysConfig.getValue(MSysConfig.PROJECT_ID_USE_CENTRALIZED_ID, "N")); // defaults to N
 		String prm_COMMENT;
-		if (!isUseCentralizedID && isUseProjectCentralizedID)
-			prm_COMMENT = MSysConfig.getValue(MSysConfig.PROJECT_ID_COMMENTS);
-		else
-			prm_COMMENT = MSysConfig.getValue(MSysConfig.DICTIONARY_ID_COMMENTS);
+		prm_COMMENT = Env.getContext(Env.getCtx(), "MigrationScriptComment");
 		if (prm_COMMENT != null && ! m_oldprm_COMMENT.equals(prm_COMMENT)) {
 			// log sysconfig comment
 			w.append("-- ");
